[func]Starting the creation of a bonus power up "Ghost"
With this bonus bird can go through obstacles during an amount of time.
To make this happen I modified the column prefab by increasing the size of the collider which triggered the score. I also created a gameobject with a ghost sprite. When the bird collides with this gameobject his trigger parameter in the collider component switches to true and his sprite becomes transparent.